remove redundant assertions
I'm removing the assertions about having 2 fasta files per barcode. Now that the code has been extended to allow for 12-plex runs, that will only have 1 barcode fasta per sample, the assertion is both violated in real cases, and is redundant because of the if statements regarding number of fastas per sample.