make slightly more verbose
I originally added these print statements when I was troubleshooting, but I think they are actually helpful information to have printed to screen every time there is a pipeline run. The output information provided allows you to see which barcode files are being pulled together for which samples, a good thing to be able to double check to make sure all data entry is correct (and corresponds to lab notebooks).