fix whitespace bug
When I wrote in the extra code chunk to handle 12-plex run scenarios, I over-indented the original 6-plex code. This meant that creating and filling the complete fasta (the file that combines both barcode fastas) was executed if the `na.fastq` condition was met. This is wrong, as we always want the complete fasta made, I just over-indented the code to make it conditional.